Revert "FishLeg Implementation Review"
