--- src/optim/FishLeg/__init__.py
@@ -1,5 +1,0 @@
-from .fishleg import FishLeg
-from .fishleg_layers import FishModel
-from .fishleg_layers import FishLinear
-
-from .fishleg_likelihood import GaussianLikelihood, BernoulliLikelihood
